Revise epub.Epub robustness, isolate info getting
Slightly improve the Epub object tolerance and proper cleaning up when
problematic epub files are parsed:
* revise __init__ so an attempt for cleaning up the temporary unzipped
folder is deleted if an Exception is raised (most likely during the
generation of EpubInfo).
* revise close() in order to delete the temporary file even if the
_zobject was not initialized properly.
* tweak _unzip() by using ZipFile.extractall if it is safe (on Python
>= 2.6.4), and adding a try..catch block if using the custom
implementation for making sure that pwd remains unchanged regardless
of potential errors.
Note that further work is still needed in order to guarantee that no
temporary files are left behind, and to refine the Exceptions thrown.

Add the 'as_model_dict()' function in order to isolate the functionality
related to getting the information that is extracted from the epub for
the purposes of building a model.