Add manager to model.Language, cleanup
Add a LanguageManager that provides a helper function for validating
a language code and return (or create) a Language model. The validation
simply consists of making sure that the 'code' exists in langlist.
Remove the save() and add() models.Language methods, in preparation for
always using the LanguageManager for creating/fetching Languages
througout the rest of the codebase.