Update and clean Books.model
Update Books.model in order to provide a base for future model
modifications, focusing on reintroducing the upload of books via the UI
and the commands, and simplyfing validation and object saving:
* Reintroduce "Book.book_file" as a FileField, in order to support
uploading via UI properly.
* Add "Book.original_path" field, as a way to provide a way of storing
the original path to the filename in order to support uploading via
command line (and returning the source file).
* Reorganize field order to improve readability.
* Update blank and null attributes of existing fields, in order to
handle None values properly and mark fields as required/not required.
* Update unique attributes (setting "file_sha256sum" as unique), as
a measure for handling duplicated entries at the DB level until a more
complex way is decided. As a result, validate_unique() has been removed
as it was redundant.
* Remove save(): it was attempting to re-set the cover image (by parsing
the Epub file) each time the object was saved, not only on creation.
* Cleanup imports.