Update addepub command
Revise 'addepub' command:
* isolate the handling of a single epub file in process_epub().
* support both specifying a single .epub file and a folder as
arguments indistinctively, allowing the user to combine both and
processing the arguments in get_epubs_paths(), that produces a clean
list of .epub paths.
* try to make sure that temporary files are cleaned up properly by
delegating on Epub for the parsing and using try..catch blocks. The
code within process_epub() has two "stages": first, try to initialize
the Epub, and if successful, proceed with creating the Book. It is
still not 100% bulletproof, and Exceptions are caught and raised in a
very coarse way.
* add some stdout information.