Add new formwizard-based view for adding books
* Add views.AddBookWizard as a contrib.formtools replacement of the
add_book view. This view uses a two-stage process: the .epub is uploaded
on the first step, validated and parsed. The information extracted from
the epub is passed onto the second step as the initial data for the user
to edit.
* Add BookUploadForm and BookMetadataForm to book.forms, as they are the
two forms used by AddBookWizard, along with the formtools app to
settings.
* Add a basic template for the view, copied from the standard formtools
template.

This change is meant to be refined in further iterations: in particular,
the template is quite raw and needs to be prettified, and some internals
of the view can be improved.