Fix download view to conform to Book.model changes
Modify the download_book() view so it returns the book_file instead of
using original_path, in order to have it working again. Added some
TODOs in order to make sure to give some thought at the implications,
and a note to remove_book() as currently the actual files are not
deleted upon model deletion.