Add custom Storage and File for symlink support
Add storage.LinkOrFileSystemStorage, a custom storage that supports
creating a symbolic link instead of copying, and also the usual copying
behaviour on the standard FileSystemStorage. The decision on whether to
use symbolic link is based on a new File subclass, LinkableFile - if
using any of the other File subclass, it behaves exactly as a
FileSystemStorage.
If for any reason (the os not supporting symbolic links, etc) the
link was unable to be created, the storage falls back to copying instead
of raising an error.