Add symlink option to addepub command
Add a new "-l"/"--link" option to addepub command. If specified, the
source files will be stored as a symbolic link instead of a copy when
adding them to the database, by making use of LinkableFile.
The default behaviour remains unchanged - if the flag is not specified,
the files will be copied as previously.