Add test for addepub command in copy mode
Add books.test.CommandAddEpubTest, for making basic black-box tests
(object creation, file creation) of the addepub command in copy mode.
The tests use the resources/epubsamples.

The tests also require the "mock" dependency (included on Python 3, but
available as a pip package on 2.x). The dependency has not been added
to requirements.txt as the tests still need some work, and the
dependency is not really needed for running the application.