Add test for addepub command in link mode
Add CommandAddEpubTest.test_addepub_symlink, replicating the existing
test_addepub_nolink but for the "symbolic link mode" of the epub
command.

Additionally, an unelegant change has been made to
LinkOrFileSystemStorage, in order to fix a problem when using mock which
made the test fail due to the "isinstance(content, LinkableFile)" check
not being correctly done (the mock object overrides the __class__
attribute). The change is only for making the test work.