Add resync command for batch-updating Book files
Add a new `resync` commands for helping the user reconcile back the
possible changes mades to the Django files on media. The command accepts
a list of items (dirs and files), and if a match is found on the DB, it
replaces Book.book_file and Book.original_path to match the new items.

By default it keeps the copied files untouched, replacing only the file
if it is a symbolic link. A "--replace-strategy" option is included for
overriding this behaviour.