Add tests for resync command, tweak storage
Add `CommandResyncTest` for testing the resync command, with a small
number of assertions on original_path and the generated files on a
single test case.
Fix an inconsistency on LinkOrFileSystemStorage, that caused deleting
symbolic links that were pointing to a broken target (such as in the
case of moving the original file) not to be deleted properly.