Django 1.6.11 upgrade: replace manage.py
Replace manage.py with a slightly tweaked version of the standard
manage.py produced by Django 1.6 "startapp" command, as the previous
one relied on the now deprecated execute_from_command_line().

A modification has been made to modify the PYTHON_PATH, appending the
parent directory before launching Django. This is meant to be a
temporary solution and might cause problems on some development
environments, and should be removed when solving issue #13 and a more
Django-like project tree is in place.