Django 1.7.11 upgrade: complete layout changes
Add book/apps.py and update settings in order to use the new django.apps
framework, remove the manage.py hack so the manage.py is now the default
django one. Update CUR_DIR variable on settings to the recommended
default, and cleanup the __init__.py on the top directory that caused
problems as django interpreted it as a module.