Django 1.7.11 upgrade: url cleanup, misc changes
Replace the quoted view names in urlpatterns with proper unquoted
views names, as they contained the wrong module path (pathagar.* should
not be used anymore, just books.*). Updated a couple of references to
the wrong module path in views and urls.
Update the tests in order to take advantage of the new default
testrunner, and the right path to resources.
Bump the django version on requirements.