Django 1.8.9 upgrade: remove django-tagging
Remove django-tagging, as it is no longer developed and it was raising
an error due to using a deprecated feature on Django 1.8. This involves
removing the tagging_tags from the templates that used them, from
settings, requirements, and adjusting the views.

The behavior of one of the views (tags()) has been temporarily modified:
it currently always returns all the tags, and not the tags for the
specified object. This should be fixed in future revisions.