Django 1.9.2 upgrade: initial migration, fixes
Provide initial migrations for the "books" app, in order to take
advantage of the migrations system and adapting to the deprecation of
the syncdb command.

Explicitely pass the initial fixture data to the tests (in particular
for the statuses). Note that the initial fixture data has not been
included as a migration, as it is likely to be revised.

Fix warning on upcoming deprecation of using render_to_string with
a context object.