Make the tests runnable again
Fix several issues that prevented the tests from being run and pass:
* revert change to validate_and_add, as it is used by the resync command
as well as the addepub command, and it needs to be able to return all
files (even if their original path seems to be on the DB) for resyncing.
* fix an issue with tags with UTF16 encoding, which prevented the
"israelsailing" book to be imported. Further changes might be needed to
deal with unicode - at the moment, no encoding/decoding is performed as
it is apparent we can't really assume that it will always be UTF8.

Plus some tiny style fixes.