Add addepub flag for skipping files, update tests
Revise "addepub" command so by default the epubs paths that match a
Book.original_file are excluded from the results automatically,
providing an "--ignore-original-path" flag to override this behavior.
Update the resync command in order to ensure those files are not
excluded, and update the tests accordingly, adding a new test to check
for the usage of that flag.