Add decorators based on Pathagar variables
Add two decorators that take into account Pathagar-defined variables.
Both decorators behave like the standard login_required decorator (and
are actually a direct copy of it) if the variable is False, and always
return the function if the variable is True:
* login_or_public_browse_required: ALLOW_PUBLIC_BROWSE
* login_or_public_add_book_required: ALLOW_PUBLIC_ADD_BOOKS

Some views and urls have been updated to use those decorators instead of
login_required, in a not comprehensive way. Further refining the rest
of the view permissions is needed.