Add preliminary permission unit test
Add test_permissions.py in order to test the permissions each kind of
user (admin, regular user, anonymous user) has for the views, depending
on the Pathagar ALLOW_PUBLIC_ADD_BOOKS and ALLOW_PUBLIC_BROWSE
variables.
This commit is meant to be improved later, as currently not all the
views are tested.