Remove add_language view
Remove add_language view, as it was unreachable from the UI and the
functionality it provided (add or change a language from the book edit
form via a browser popup) can now be provided using the m2m mechanisms.

In turn, this triggered quite a cleanup other than views and urls: in
particular, removing popuphandler/selectwith pop along with its
templates, plus removing RelatedObjectLookups.js (which provided the
popup handling functionality in the browser) and references to it.