Fix OPDS feed author and publisher handling
Update OPDS feeds to they do not crash due to the model changes in
author and publisher. This change only ensures that valid values are
getting passed to handler.add_quick_element and write_person_construct,
and more testing and diagnosing should be performed as it is not
guaranteed that they fix all the possible OPDS issues.