Review book edit form and template
Revise book_form.html in order to make better use of bootstrap features
and prepare it for being able to use it also with the UI add book with
wizard feature.
Modify BookEditForm taking advantage of Meta for including just the
desired fields and set the labels. Remove the save() method as book_file
cannot be modified using the form, save_m2m() is automatically called
and the deletion of cover_img is now handled via a custom ImageField.
Add a custom template tag for saving some typing and complexity on the
template when setting the field widths.