Cleanup Epub.as_model_dict changes
Move the authors and the publishers back into the dict returned by
Epub.as_model_dict(), as that is the purpose of the function, which acts
as a small adapter for making both the addepub and the import via UI
view receive that data.
Updated the Wizard view to take into account this information and
converting it for the initial values of the form on step 2, and cleaned
up addepub cruft.