Share template between AddBookWizard, BookEditView
Split the original "book_form" into an additional snippet form, which is
included from the book edit and from the add book wizard templates, so
they are rendered in the same way.
Changed BookMetadataForm (used on step 2) to inherit from BookEditForm
directly, in order to reuse the m2m fields.