Reorganize base template footer
Tweak base template so the footer is now a "footer" html element, and it
is always displayed (except on the login page). Change the sticky-footer
css file to make it sticky, and added some spacing/color rules.
Update the templates that were using the block (all of them with empty
content) so they use the standard header.
The inner contents of the footer have not been modified, further changes
are needed.