Change RSS link on base template depending on view
Reintroduce the feature where, for those views that display a list and
use the "list_by" parameter, the RSS link points to the corresponding
feed instead of pointing always to the "latest" feed.
This is done via a custom template tag - the module has been renamed to
a more general "pathagar_common" name.