Remove book_list.html, move accounts templates
Remove "book_list.html" as it was not used, and renamed
"book_list2.html" to "book_list.html", updating a view in the process.
Move accounts/templates/accounts to templates/, to make it follow the
same pattern as books templates.
Fix problem on tag_list that caused an error when running tests.